+++ changes.md
@@ -1,5 +1,7 @@
 * leaflet-elevation-2.2.8 with own presentation of results.
-* compability with themes and (block)editor
 * new option in markerClusterGroup: visible
-* fixed bug zoomhomemap, hover, multielevation
-* sgpx does not work with some themes (like TT2)
+* fixed some bugs in zoomhomemap, hover, multielevation
+* compability with themes and (block)editor
+* sgpx (from WP GPX Maps) does not work with some themes (like TT2)
+* new shortcode: choropleth - https://github.com/timwis/leaflet-choropleth
+* new shortcodes leaflet-optiongroups and leaflet-featuregroups (Leaflet.FeatureGroup.SubGroup) (todo: documentation)